refactor: artistId 타입 Long으로 변경
change artistId type from Integer to Long across related classes

+++ backend/BE/src/main/java/com/a404/duckonback/dto/SignupRequestDTO.java
@@ -16,6 +16,6 @@ public class SignupRequestDTO {
     private String password;
     private String nickname;
     private String language;
-    private List<Integer> artistList;
+    private List<Long> artistList;
     private MultipartFile profileImg;
 }
